Moved Announcements and Clients to admin directory, and modified their tests to work in new location
