+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/clients/ClientsController.java
@@ -431,7 +431,7 @@ public class ClientsController extends BaseController {
 	public String removeClient(@RequestParam Long id) {
 		clientDetailsService.delete(id);
 
-		return "redirect:/clients";
+		return "redirect:/admin/clients";
 	}
 
 	/**
