+++ src/main/webapp/resources/js/pages/admin/components/Admin.jsx
@@ -24,10 +24,10 @@ import { setBaseUrl } from "../../../utilities/url-utilities";
 const AdminStatistics = lazy(() => import("./AdminStatistics"));
 const AdminUsersPage = lazy(() => import("./AdminUsersPage"));
 const AdminUserGroupsPage = lazy(() => import("../../UserGroupsPage/components/UserGroupsPage"));
-const ClientsPage = lazy(() => import("../../clients/ClientsPage"));
+const ClientListingPage = lazy(() => import("./clients/listing/ClientListingPage"));
 const AdminRemoteApiPage = lazy(() => import("./AdminRemoteApiPage"));
 const AdminSequencingRunsPage = lazy(() => import("./AdminSequencingRunsPage"));
-const AnnouncementAdminPage = lazy(() => import("../../announcement/AnnouncementAdminPage"));
+const AnnouncementAdminPage = lazy(() => import("./announcements/AnnouncementAdminPage"));
 
 export default function Admin() {
   const DEFAULT_URL = setBaseUrl("/admin");
@@ -60,7 +60,7 @@ export default function Admin() {
                     `${DEFAULT_URL}/${ADMIN.USERGROUPS}`
                   }
                 />
-                <ClientsPage
+                <ClientListingPage
                   path={
                     `${DEFAULT_URL}/${ADMIN.CLIENTS}`
                   }
