+++ src/main/webapp/resources/js/pages/admin/components/announcements/AnnouncementAdminPage.jsx
@@ -7,12 +7,12 @@ import {
   createNewAnnouncement,
   deleteAnnouncement,
   updateAnnouncement
-} from "../../apis/announcements/announcements";
+} from "../../../../apis/announcements/announcements";
 import { CreateNewAnnouncement } from "./CreateNewAnnouncement";
-import { setBaseUrl } from "../../utilities/url-utilities";
-import { PageWrapper } from "../../components/page/PageWrapper";
+import { setBaseUrl } from "../../../../utilities/url-utilities";
+import { PageWrapper } from "../../../../components/page/PageWrapper";
 import { AnnouncementsTable } from "./AnnouncementsTable";
-import { PagedTableProvider } from "../../components/ant.design/PagedTable";
+import { PagedTableProvider } from "../../../../components/ant.design/PagedTable";
 
 /**
  * React component to render the administration announcements page.
